test(upload): 13a commit D — the chunked-upload battery (26 cases)
Resolver: tier defaults (1GB PAID / 20GB ADMIN from env), per-user
override BEATS tier (the add-on infra, exercised end to end via a
SQL flip like the future flip-kit will do), unknown-role
conservative, the §1 friendly wording.

Init: happy path (chunk plan returned, staging created), extension
allowlist, over-tier 413 + wording (fail-fast before bytes), admin
tier allows bigger, one-active-session rule, quota-headroom 413 with
the honest math (~500MB-headroom seed so the QUOTA check is what
fires, not the file cap), not-your-section 403.

Chunks: append + IDEMPOTENT re-send (the network-retry contract),
index bounds, wrong-size chunk (lying client), another user's
session → 404 (no existence leak), swept session → 409 (§3a's
explicit race answer).

Complete: end-to-end proof (init → chunks → complete → Video row
'queued', transcribe job REGISTERED at creation (the 9/19 lesson),
staging dir consumed, final file exists at declared size), missing
chunks 400 with the missing list.

Sweeper: TTL honored (old swept: dir gone + row cancelled; fresh
kept), idempotent double-pass (the c827a7b claim-durability shape
applied to this sweeper).

Quota interactions: staging counts at DECLARED size (reservation
model), DELETE releases instantly, completion converts the
reservation into the Video's file_size (meter number unchanged,
class changes), last-cancelled banner feed.

Two service fixes the battery forced: chunk_size int() (float MB
settings crashed byte math), and the battery itself validated the
last-chunk-remainder semantics (declared - (n-1)*csize) via the
_chunk_payload helper.

Full suite: 1561 passed.

+++ app/services/upload_sessions.py
@@ -135,7 +135,7 @@ def create_session(
     # more actionable error wins).
     check_quota_headroom(db, user, declared_size)
 
-    chunk_size = settings.upload_chunk_size_mb * 1024 * 1024
+    chunk_size = int(settings.upload_chunk_size_mb * 1024 * 1024)
     total_chunks = math.ceil(declared_size / chunk_size)
 
     session = UploadSession(
